+++ pcsx2/R3000A.cpp
@@ -205,7 +205,7 @@ __ri void iopEventTest(void)
 	{
 		// start the next branch at the next counter event by default
 		// the interrupt code below will assign nearer branches if needed.
-		if (psxNextDeltaCounter < (psxRegs.iopNextEventCycle - psxNextStartCounter))
+		if (psxNextDeltaCounter < static_cast<s32>(psxRegs.iopNextEventCycle - psxNextStartCounter))
 			psxRegs.iopNextEventCycle = psxNextStartCounter + psxNextDeltaCounter;
 	}
